requirements.txt
# Datadog APM with Exception Replay (requires >= 1.10.0 for exception replay)
ddtrace>=2.0.0
datadog==0.47.0

# Environment Configuration
python-dotenv==1.0.0
